feat: core backend functionality complete

+++ backend/src/routes/message.js
@@ -44,35 +44,32 @@ router.post(
           res.json({ result: "You are not signed in." });
         } else {
           const fullVerify = async () => {
-            if (req.body.users.includes(authData.user.username)) {
-              const acc = await req.context.models.Messenger.findOne({
-                username: authData.user.username,
-                password: authData.user.password,
+            const acc = await req.context.models.Messenger.findOne({
+              username: authData.user.username,
+              password: authData.user.password,
+            });
+            const findChat = await req.context.models.Chat.findById(
+              req.params.chatId,
+            );
+            if (acc && findChat && findChat.users.includes(acc._id)) {
+              const message = await req.context.models.Chatmessage.create({
+                text: req.body.message,
+                user: acc._id,
+                date: Date.now(),
+              }).catch((error) => {
+                error.statusCode = 400;
+                next(error);
               });
-              if (acc && acc.users.includes(authData.user.username)) {
-                async () => {
-                  const message = await req.context.models.Chatmessage.create({
-                    text: req.body.text,
-                    user: authData.user.username,
-                    date: Date.now(),
-                  }).catch((error) => {
-                    error.statusCode = 400;
-                    next(error);
-                  });
-                  const chat = await req.context.models.Chat.findByIdAndUpdate(
-                    req.params.chatId,
-                    {
-                      $push: { messages: message._id },
-                      lastMessage: message._id,
-                    },
-                  );
-                  return res.json({
-                    result: "Message posted",
-                  });
-                };
-              } else {
-                res.json({ result: "Invalid authentication token" });
-              }
+              const chat = await req.context.models.Chat.findByIdAndUpdate(
+                req.params.chatId,
+                {
+                  $push: { messages: message._id },
+                  lastMessage: message._id,
+                },
+              );
+              return res.json({
+                result: "Message posted",
+              });
             } else {
               res.json({ result: "Invalid authentication token" });
             }
@@ -98,31 +95,25 @@ router.put(
           res.json({ result: "You are not signed in." });
         } else {
           const fullVerify = async () => {
+            const acc = await req.context.models.Messenger.findOne({
+              username: authData.user.username,
+              password: authData.user.password,
+            });
             const message = await req.context.models.Chatmessage.findById(
               req.params.messageId,
             ).catch((error) => {
               error.statusCode = 400;
               next(error);
             });
-            if (message.user === authData.user.username) {
-              const acc = await req.context.models.Messenger.findOne({
-                username: authData.user.username,
-                password: authData.user.password,
+            if (acc && message.user.toString() === acc.id) {
+              const update =
+                await req.context.models.Chatmessage.findByIdAndUpdate(
+                  req.params.messageId,
+                  { text: req.body.message },
+                );
+              return res.json({
+                result: "Message updated",
               });
-              if (acc) {
-                async () => {
-                  const update =
-                    await req.context.models.Chatmessage.findByIdAndUpdate(
-                      req.params.messageId,
-                      { text: req.body.message },
-                    );
-                  return res.json({
-                    result: "Message updated",
-                  });
-                };
-              } else {
-                res.json({ result: "Invalid authentication token" });
-              }
             } else {
               res.json({ result: "Invalid authentication token" });
             }
@@ -146,36 +137,49 @@ router.delete("/:chatId/:messageId", verifyToken, async (req, res, next) => {
           error.statusCode = 400;
           next(error);
         });
-        if (message.user === authData.user.username) {
-          const acc = await req.context.models.Messenger.findOne({
-            username: authData.user.username,
-            password: authData.user.password,
-          });
-          if (acc) {
-            async () => {
-              const chat = await req.context.models.Chat.findById(
-                req.params.chatId,
-              );
-              const del =
-                await req.context.models.Chatmessage.findByIdAndDelete(
-                  req.params.messageId,
-                );
-              if (chat.lastMessage === message._id) {
-                async () => {
-                  const delChat =
-                    await req.context.models.Chat.findByIdAndDelete(
-                      req.params.chatId,
-                    );
-                  return res.json({ result: "Chat deleted" });
-                };
-              } else {
-                return res.json({
-                  result: "Message deleted",
+        const acc = await req.context.models.Messenger.findOne({
+          username: authData.user.username,
+          password: authData.user.password,
+        });
+        if (message && acc && message.user.toString() === acc.id) {
+          const chat = await req.context.models.Chat.findById(
+            req.params.chatId,
+          );
+          const del = await req.context.models.Chatmessage.findByIdAndDelete(
+            req.params.messageId,
+          );
+
+          if (chat.messages.length === 1) {
+            chat.users.forEach(async (id) => {
+              const eachUser =
+                await req.context.models.Messenger.findByIdAndUpdate(id, {
+                  $pull: { chats: chat._id },
+                }).catch((err) => {
+                  res.send(err);
                 });
-              }
-            };
+            });
+            const delChat = await req.context.models.Chat.findByIdAndDelete(
+              req.params.chatId,
+            );
+
+            return res.json({ result: "Chat deleted" });
           } else {
-            res.json({ result: "Invalid authentication token" });
+            const sortedChat = await req.context.models.Chat.findById(
+              req.params.chatId,
+            )
+              .sort({ date: -1 })
+              .limit(1);
+            const updatedChat = await req.context.models.Chat.findByIdAndUpdate(
+              req.params.chatId,
+              {
+                lastMessage: sortedChat.messages[0]._id,
+                $pull: { messages: req.params.messageId },
+              },
+            );
+
+            return res.json({
+              result: "Message deleted",
+            });
           }
         } else {
           res.json({ result: "Invalid authentication token" });
